+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Fixes
-
+ - Music playing
+ - Commands working
  - Include Frappé correctly.